restructured Plugin interface, all Plugin interfaces implement one common OpenDiabetes Plugin interface now --> preventing duplicate code
