+++ src/de/opendiabetes/vault/plugin/common/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * //TODO javadoc
+ */
+package de.opendiabetes.vault.plugin.common;
