+++ src/de/opendiabetes/vault/plugin/importer/Importer.java
@@ -18,6 +18,7 @@ package de.opendiabetes.vault.plugin.importer;
 
 import de.opendiabetes.vault.container.RawEntry;
 import de.opendiabetes.vault.container.VaultEntry;
+import de.opendiabetes.vault.plugin.common.OpenDiabetesPlugin;
 import org.pf4j.ExtensionPoint;
 
 import java.util.List;
@@ -32,13 +33,7 @@ import java.util.logging.Logger;
  * It also serves as the {@link org.pf4j.ExtensionPoint} where the plugins hook up.
  * Therefore all importer plugins must implement this interface to get recognized as importer.
  */
-public interface Importer extends ExtensionPoint {
-
-    /**
-     * Logger object of all Importers. Logs all messages of the importers to a human readable file.
-     */
-    Logger LOG = Logger.getLogger(Importer.class.getName());
-
+public interface Importer extends ExtensionPoint, OpenDiabetesPlugin {
 
     /**
      * Getter for the importFilePath.
@@ -84,34 +79,4 @@ public interface Importer extends ExtensionPoint {
      * will return empty lists afterwards.
      */
     void clearData();
-
-    /**
-     * Method to load the plugin's configuration file.
-     *
-     * @param configuration the configuration object
-     * @return True if configuration can be loaded, false otherwise.
-     */
-    boolean loadConfiguration(Properties configuration);
-
-    /**
-     * Method to register listeners to the Plugins.
-     * The GUI for example can implement onStatusCallback behavior and register its interface here to get notified by a status update.
-     *
-     * @param listener A listener.
-     */
-    void registerStatusCallback(StatusListener listener);
-
-    /**
-     * Interface which defines the methods called on a status update.
-     * Must be implemented by any listener of this plugin to handle the status update.
-     */
-    interface StatusListener {
-        /**
-         * Is called multiple times on all listeners during the import process to notify them about the import progress.
-         *
-         * @param progress Percentage of completion.
-         * @param status   Current Status.
-         */
-        void onStatusCallback(int progress, String status);
-    }
 }
